+++ templates/antimicrobial_resistance.sh
@@ -26,22 +26,22 @@ if [[ !{params.compress} == "true" ]]; then
 fi
 
 tar -xzvf !{amrdb}
-mkdir antimicrobial_resistance
+mkdir !{amrdir}
 
 # amrfinder Version
 echo "# amrfinder Version" >> ${LOG_DIR}/!{task.process}.versions
 amrfinder --version >> ${LOG_DIR}/!{task.process}.versions 2>&1
 amrfinder -n !{sample}.ffn \
-          -d amrdb/ \
-          -o antimicrobial_resistance/!{sample}-gene-report.txt \
+          -d amrfinderdb/ \
+          -o !{amrdir}/!{sample}-gene-report.txt \
           --ident_min !{params.amr_ident_min} \
           --coverage_min !{params.amr_coverage_min} \
           --translation_table !{params.amr_translation_table} \
           --threads !{task.cpus} !{organism_gene} !{plus} !{report_common} > ${LOG_DIR}/amrfinder-gene.out 2> ${LOG_DIR}/amrfinder-gene.err
 
 amrfinder -p !{sample}.faa \
-          -d amrdb/ \
-          -o antimicrobial_resistance/!{sample}-protein-report.txt \
+          -d amrfinderdb/ \
+          -o !{amrdir}/!{sample}-protein-report.txt \
           --ident_min !{params.amr_ident_min} \
           --coverage_min !{params.amr_coverage_min} \
           --translation_table !{params.amr_translation_table} \